tds_convert_real: Detect SYB(U)INT4 overflow precisely.
Cast IS_(U)INT's argument to double, per Clang's warnings that it will
otherwise increase each upper bound by one in the course of implicitly
converting it to float.

git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@93915 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/dbapi/driver/ftds100/freetds/tds/convert.c
@@ -1531,13 +1531,13 @@ tds_convert_real(const TDS_REAL* src, int desttype, CONV_RESULT * cr)
 		return sizeof(TDS_USMALLINT);
 		break;
 	case SYBINT4:
-		if (!IS_INT(the_value))
+		if (!IS_INT((double)the_value))
 			return TDS_CONVERT_OVERFLOW;
 		cr->i = (TDS_INT) the_value;
 		return sizeof(TDS_INT);
 		break;
 	case SYBUINT4:
-		if (!IS_UINT(the_value))
+		if (!IS_UINT((double)the_value))
 			return TDS_CONVERT_OVERFLOW;
 		cr->ui = (TDS_UINT) the_value;
 		return sizeof(TDS_UINT);
